MMDデフォルトモデルを追加し、スフィアマップを追加し、ピクセル透過処理を実装

+++ IQByDirectX12/IQByDirectX12/Source/CommandQueue.cpp
@@ -71,7 +71,7 @@ void CommandQueue::ExecuteCommandList(const std::vector<std::shared_ptr<Graphics
 	{
 		commandListsArray[i] = commandLists[i]->GetCommandList().Get();
 	}
-	mCommandQueue->ExecuteCommandLists(commandListsArray.size(), commandListsArray.data());
+	mCommandQueue->ExecuteCommandLists(static_cast<UINT>(commandListsArray.size()), commandListsArray.data());
 }
 
 void CommandQueue::Signal()
